feeling-from-photo-API/requirements.txt
Flask==3.0.3
pillow==10.3.0
transformers==4.41.2
